fix: use dynamic colors for overlays in TagDetailPage and UserDetailPage
- Replace primaryText overlay with dynamic(light: .black, dark: .white) to maintain proper darkening effect in light mode and lightening in dark mode
- Fixes build issues by moving SettingActions into SettingState temporarily
- Restores original overlay opacity values (0.3/0.1)
