+++ V2er/State/DataFlow/Reducers/SettingReducer.swift
@@ -9,7 +9,20 @@
 import Foundation
 
 func settingStateReducer(_ state: SettingState, _ action: Action) -> (SettingState, Action?) {
-    return (SettingState(), nil)
+    var state = state
+    var followingAction: Action? = action
+    
+    switch action {
+    case let action as SettingActions.ChangeAppearanceAction:
+        state.appearance = action.appearance
+        // Save to UserDefaults
+        UserDefaults.standard.set(action.appearance.rawValue, forKey: "appearanceMode")
+        followingAction = nil
+    default:
+        break
+    }
+    
+    return (state, followingAction)
 }
